# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: LMS Finance & Voucher Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /students:
    get:
      operationId: listStudents
      tags: [students]
      summary: List all students
      parameters:
        - name: courseId
          in: query
          schema:
            type: integer
        - name: course
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive]
        - name: search
          in: query
          schema:
            type: string
            description: Search by name, studentCode, or phone
      responses:
        "200":
          description: List of students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"
    post:
      operationId: createStudent
      tags: [students]
      summary: Create a student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStudentInput"
      responses:
        "201":
          description: Student created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /students/{id}:
    get:
      operationId: getStudent
      tags: [students]
      summary: Get a student by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateStudent
      tags: [students]
      summary: Update a student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStudentInput"
      responses:
        "200":
          description: Updated student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteStudent
      tags: [students]
      summary: Delete a student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /students/{id}/ledger:
    get:
      operationId: getStudentLedger
      tags: [students]
      summary: Get complete financial ledger for a student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student ledger
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentLedger"

  # ── Fee Structures (legacy) ─────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error, message]

    # ── Course ────────────────────────────────────────────────────────────────
...
    Student:
      type: object
      properties:
        id:
          type: integer
        studentCode:
          type: string
          description: Auto-generated 6-digit code e.g. 100001
        name:
          type: string
        course:
          type: string
          description: Course name (denormalized)
        courseId:
          type: integer
        fatherName:
          type: string
        phone:
          type: string
        address:
          type: string
        status:
          type: string
          enum: [active, inactive]
        enrollmentDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
          description: Auto-calculated from enrollmentDate + course.durationMonths
        discountAmount:
          type: number
          description: Monthly discount applied to this student
        effectiveFee:
          type: number
          description: monthlyFee - discountAmount
        createdAt:
          type: string
          format: date-time
      required: [id, studentCode, name, course, courseId, status, enrollmentDate, createdAt]
    CreateStudentInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        courseId:
          type: integer
          description: ID of the course from /courses
        fatherName:
          type: string
        phone:
          type: string
        address:
          type: string
        status:
          type: string
          enum: [active, inactive]
          default: active
        enrollmentDate:
          type: string
          format: date
        discountAmount:
          type: number
          minimum: 0
          default: 0
      required: [name, courseId, enrollmentDate]
    UpdateStudentInput:
      type: object
      properties:
        name:
          type: string
        courseId:
          type: integer
        fatherName:
          type: string
        phone:
          type: string
        address:
          type: string
        status:
          type: string
          enum: [active, inactive]
        enrollmentDate:
          type: string
          format: date
        discountAmount:
          type: number
          minimum: 0
    StudentLedger:
      type: object
      properties:
        student:
          $ref: "#/components/schemas/Student"
        entries:
          type: array
          items:
            $ref: "#/components/schemas/LedgerEntry"
        totalFee:
          type: number
        totalReceived:
          type: number
        totalPending:
          type: number
      required: [student, entries, totalFee, totalReceived, totalPending]
    LedgerEntry:
      type: object
      properties:
        voucherId:
          type: integer
        month:
          type: integer
        year:
          type: integer
        monthName:
          type: string
        fee:
          type: number
        received:
          type: number
        pending:
          type: number
        status:
          type: string
          enum: [unpaid, partial, paid]
      required: [voucherId, month, year, monthName, fee, received, pending, status]

    # ── Fee Structure (legacy) ───────────────────────────────────────────────
...